#1- 개행단위 입력받기
