+++ Algorithm/Algorithm/#1-Practice/main.swift
@@ -1,5 +1,5 @@
 //
-//  main.swift
+//  여러가지 입력받기.swift
 //  Algorithm
 //
 //  Created by 김인영 on 2023/08/31.
@@ -7,18 +7,7 @@
 
 import Foundation
 
-let input = readLine()
-print(type(of: input))
+// 개행 단위 입력받기
 
-let inputOut = input!
-print(type(of: inputOut))
-
-let intVar = Int(inputOut)
-print(type(of: intVar))
-
-let result = intVar!
-print(type(of: result))
-
-// 한 줄 요약
-let newInput = Int(readLine()!)!
-print(newInput)
+let input = readLine()!.split(separator: " ")
+print(input)
