+++ Algorithm/Algorithm/#1-Practice/정수형_입력받기.swift
@@ -1,0 +1,24 @@
+//
+//  main.swift
+//  Algorithm
+//
+//  Created by 김인영 on 2023/08/31.
+//
+
+import Foundation
+
+let input = readLine()
+print(type(of: input))
+
+let inputOut = input!
+print(type(of: inputOut))
+
+let intVar = Int(inputOut)
+print(type(of: intVar))
+
+let result = intVar!
+print(type(of: result))
+
+// 한 줄 요약
+let newInput = Int(readLine()!)!
+print(newInput)
